+++ packages/cli-kit/src/public/node/cli.ts
@@ -1,4 +1,5 @@
 import {isTruthy} from './context/utilities.js'
+import {launchCLI as defaultLaunchCli} from './cli-launcher.js'
 import {cacheClear} from '../../private/node/conf-store.js'
 import {environmentVariables} from '../../private/node/constants.js'
 import {Flags} from '@oclif/core'
@@ -74,7 +75,7 @@ function forceNoColor(argv: string[] = process.argv, env: NodeJS.ProcessEnv = pr
  */
 export async function runCLI(
   options: RunCLIOptions & {runInCreateMode?: boolean},
-  launchCLI: (options: {moduleURL: string}) => Promise<void>,
+  launchCLI: (options: {moduleURL: string}) => Promise<void> = defaultLaunchCli,
   argv: string[] = process.argv,
   env: NodeJS.ProcessEnv = process.env,
   versions: NodeJS.ProcessVersions = process.versions,
@@ -111,7 +112,7 @@ async function addInitToArgvWhenRunningCreateCLI(
  */
 export async function runCreateCLI(
   options: RunCLIOptions,
-  launchCLI: (options: {moduleURL: string}) => Promise<void>,
+  launchCLI: (options: {moduleURL: string}) => Promise<void> = defaultLaunchCli,
   argv: string[] = process.argv,
   env: NodeJS.ProcessEnv = process.env,
   versions: NodeJS.ProcessVersions = process.versions,