+++ packages/theme/src/cli/commands/theme/dev.ts
@@ -4,6 +4,7 @@ import ThemeCommand, {FlagValues} from '../../utilities/theme-command.js'
 import {dev} from '../../services/dev.js'
 import {DevelopmentThemeManager} from '../../utilities/development-theme-manager.js'
 import {findOrSelectTheme} from '../../utilities/theme-selector.js'
+import {metafieldsPull} from '../../services/metafields-pull.js'
 import {Flags} from '@oclif/core'
 import {globalFlags} from '@shopify/cli-kit/node/cli'
 import {Theme} from '@shopify/cli-kit/node/themes/types'
@@ -153,6 +154,16 @@ You can run this command only in a directory that matches the [default Shopify t
       ignore,
       only,
       notify: flags.notify,
+    })
+
+    await metafieldsPull({
+      path: flags.path,
+      password: flags.password,
+      store: flags.store,
+      force: flags.force,
+      verbose: flags.verbose,
+      noColor: flags['no-color'],
+      silent: true,
     })
   }
 }