+++ packages/theme/src/index.ts
@@ -9,6 +9,7 @@ import ListCommnd from './cli/commands/theme/list.js'
 import Open from './cli/commands/theme/open.js'
 import Package from './cli/commands/theme/package.js'
 import Publish from './cli/commands/theme/publish.js'
+import MetafieldsPull from './cli/commands/theme/metafields/pull.js'
 import Pull from './cli/commands/theme/pull.js'
 import Push from './cli/commands/theme/push.js'
 import Rename from './cli/commands/theme/rename.js'
@@ -24,6 +25,7 @@ const COMMANDS = {
   'theme:info': ThemeInfo,
   'theme:language-server': LanguageServer,
   'theme:list': ListCommnd,
+  'theme:metafields:pull': MetafieldsPull,
   'theme:open': Open,
   'theme:package': Package,
   'theme:publish': Publish,