+++ packages/plugin-did-you-mean/CHANGELOG.md
@@ -1,5 +1,44 @@
 # @shopify/plugin-did-you-mean
 
+## 3.72.0
+
+### Patch Changes
+
+- @shopify/cli-kit@3.72.0
+
+## 3.71.5
+
+### Patch Changes
+
+- Updated dependencies [94a7d4ce4f]
+  - @shopify/cli-kit@3.71.5
+
+## 3.71.4
+
+### Patch Changes
+
+- Updated dependencies [94953a671e]
+  - @shopify/cli-kit@3.71.4
+
+## 3.71.3
+
+### Patch Changes
+
+- @shopify/cli-kit@3.71.3
+
+## 3.71.2
+
+### Patch Changes
+
+- Updated dependencies [a395820e7b]
+  - @shopify/cli-kit@3.71.2
+
+## 3.71.1
+
+### Patch Changes
+
+- @shopify/cli-kit@3.71.1
+
 ## 3.71.0
 
 ### Patch Changes